requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
asyncpg==0.31.0
axios==0.4.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
docker==7.1.0
exceptiongroup==1.3.1
fastapi==0.135.1
fonttools==4.62.1
greenlet==3.3.2
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.5.0
langchain-core==1.2.19
langchain-groq==1.1.2
langgraph==1.1.2
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.17
lxml==6.0.2
markdown-it-py==4.0.0
matplotlib==3.10.8
mdurl==0.1.2
numpy==2.2.6
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
patsy==1.0.2
pillow==12.1.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
pywin32==311
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
sqlmodel==0.0.37
starlette==0.52.1
statsmodels==0.14.6
tenacity==9.1.4
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
xxhash==3.6.0
zstandard==0.25.0
